// src/interfaces/complete-api-params.interface.d.ts
export interface ICompleteApiParams {
    click_trans_id: number
    service_id: number
    click_paydoc_id: number
    merchant_trans_id: number
    amount: number
    action: number
    error: number
    error_note: string
    sign_time: string
    sign_string: string
    merchant_prepare_id: number
}